+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/IMovimientoDineroService.java
@@ -6,9 +6,9 @@
 
  public interface IMovimientoDineroService {
 
-     public MovimientoDinero findById(int id);
+     public MovimientoDinero findById(long id);
      public List<MovimientoDinero> findAll();
      public MovimientoDinero creatMovimiento(MovimientoDinero movimiento);
      public MovimientoDinero updateMovimiento(int id, MovimientoDinero movimiento);
-     public void deletMovimiento(int id);
+     public void deletMovimiento(long id);
  }
