+++ ingresos_egresos/src/main/java/servicios/exequiales/ingresosyegresos/ingresos_egresos/Service/MovimientoDineroService.java
@@ -5,7 +5,7 @@
  import org.springframework.stereotype.Service;
  import servicios.exequiales.ingresosyegresos.ingresos_egresos.repository.IMovimientoRepository;
 
-import java.util.ArrayList;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -15,19 +15,16 @@ import java.util.Optional;
      @Autowired
      private IMovimientoRepository movimientoRepository;
 
-
      @Override
-     public MovimientoDinero findById(int id) {
+     public MovimientoDinero findById(long id) {
          Optional<MovimientoDinero> movimientoDinero = movimientoRepository.findById((long) id);
          return movimientoDinero.get();
      }
-
      @Override
      public List<MovimientoDinero> findAll() {
          List<MovimientoDinero> movimientos = (List<MovimientoDinero>) movimientoRepository.findAll();
          return movimientos;
      }
-
      @Override
      public MovimientoDinero creatMovimiento(MovimientoDinero movimiento)  {
          return movimientoRepository.save(movimiento);
@@ -38,7 +35,7 @@ import java.util.Optional;
          return movimientoRepository.save(movimiento);
      }
      @Override
-     public void deletMovimiento(int id)  {
+     public void deletMovimiento(long id)  {
          movimientoRepository.deleteById((long)id);
      }
  }
